+++ src/components/AvatarStack/__tests__/stories.ts
@@ -1,0 +1,5 @@
+import {composeStories} from '@storybook/react';
+
+import * as CSFStories from '../__stories__/AvatarStack.stories';
+
+export const AvatarStackStories = composeStories(CSFStories);