+++ src/components/Sheet/__tests__/helpersPlaywright.tsx
@@ -1,0 +1,5 @@
+import {composeStories} from '@storybook/react';
+
+import * as stories from '../__stories__/DefaultShowcase/DefaultShowcase.stories';
+
+export const SheetStories = composeStories(stories);