+++ src/components/index.ts
@@ -19,6 +19,7 @@ export * from './Divider';
 export * from './DropdownMenu';
 export * from './Hotkey';
 export * from './Icon';
+export * from './AvatarStack';
 export * from './Label';
 export * from './Link';
 export * from './List';